Larger object id field
This is required to process the twitter data set as there are too many items for
a single unsigned int.

+++ src/common/DataObject.hpp
@@ -16,7 +16,7 @@
 class DataObject
 {
 	public:
-		using Id = unsigned int;
+		using Id = unsigned long long;
 
 		/**
 		 * Create a new spatial object.
